Refactor cli tools and update readme
